Trait greenglas::modifier::Modifier

pub trait Modifier<F>
where
    F: ?Sized,
{
    // Required method
    fn modify(self, _: &mut F);
}

Allows use of the implemented type as an argument to Set::set.

This allows types to be used for ad-hoc overloading of Set::set to perform complex updates to the parameter of Modifier.

Required Methods

fn modify(self, _: &mut F)

Modify F with self.