Trait greenglas::Set

pub trait Set {
    pub fn set<M>(self, modifier: M) -> Self
    where
        M: Modifier<Self>,
    { ... }

    pub fn set_mut<M>(&mut self, modifier: M) -> &mut Self
    where
        M: Modifier<Self>,
    { ... }
}

A trait providing the set and set_mut methods for all types.

Simply implement this for your types and they can be used with modifiers.

Provided methods

pub fn set<M>(self, modifier: M) -> Self where
M: Modifier<Self>,

Modify self using the provided modifier.

pub fn set_mut<M>(&mut self, modifier: M) -> &mut Self where
M: Modifier<Self>,

Modify self through a mutable reference with the provided modifier.

Loading content...